+++ test/ReaderTask.ts
@@ -15,24 +15,20 @@ describe('ReaderTask', () => {
 
   it('map', async () => {
     const double = (n: number): number => n * 2
-    const x = await pipe(_.of(1), _.map(double))({})()
-    assert.deepStrictEqual(x, 2)
+    assert.deepStrictEqual(await pipe(_.of(1), _.map(double))({})(), 2)
   })
 
   it('ap', async () => {
     const double = (n: number): number => n * 2
-    const x = await pipe(_.of(double), _.ap(_.of(1)))({})()
-    assert.deepStrictEqual(x, 2)
+    assert.deepStrictEqual(await pipe(_.of(double), _.ap(_.of(1)))({})(), 2)
   })
 
   it('apFirst', async () => {
-    const x = await pipe(_.of('a'), _.apFirst(_.of('b')))({})()
-    assert.deepStrictEqual(x, 'a')
+    assert.deepStrictEqual(await pipe(_.of('a'), _.apFirst(_.of('b')))({})(), 'a')
   })
 
   it('apSecond', async () => {
-    const x = await pipe(_.of('a'), _.apSecond(_.of('b')))({})()
-    assert.deepStrictEqual(x, 'b')
+    assert.deepStrictEqual(await pipe(_.of('a'), _.apSecond(_.of('b')))({})(), 'b')
   })
 
   it('chain', async () => {
@@ -43,23 +39,19 @@ describe('ReaderTask', () => {
 
   it('chainFirst', async () => {
     const f = (a: string) => _.of(a.length)
-    const x = await pipe(_.of('foo'), _.chainFirst(f))({})()
-    assert.deepStrictEqual(x, 'foo')
+    assert.deepStrictEqual(await pipe(_.of('foo'), _.chainFirst(f))({})(), 'foo')
   })
 
   it('flatten', async () => {
-    const x = await pipe(_.of(_.of('a')), _.flatten)({})()
-    assert.deepStrictEqual(x, 'a')
+    assert.deepStrictEqual(await pipe(_.of(_.of('a')), _.flatten)({})(), 'a')
   })
 
   it('of', async () => {
-    const e = await _.fromReader(R.of(1))({})()
-    assert.deepStrictEqual(e, 1)
+    assert.deepStrictEqual(await _.fromReader(R.of(1))({})(), 1)
   })
 
   it('fromIO', async () => {
-    const e = await _.fromIO(() => 1)({})()
-    assert.deepStrictEqual(e, 1)
+    assert.deepStrictEqual(await _.fromIO(() => 1)({})(), 1)
   })
 
   // -------------------------------------------------------------------------------------
@@ -67,23 +59,19 @@ describe('ReaderTask', () => {
   // -------------------------------------------------------------------------------------
 
   it('ask', async () => {
-    const e = await _.ask<number>()(1)()
-    return assert.deepStrictEqual(e, 1)
+    return assert.deepStrictEqual(await _.ask<number>()(1)(), 1)
   })
 
   it('asks', async () => {
-    const e = await _.asks((s: string) => s.length)('foo')()
-    return assert.deepStrictEqual(e, 3)
+    return assert.deepStrictEqual(await _.asks((s: string) => s.length)('foo')(), 3)
   })
 
   it('fromTask', async () => {
-    const e = await _.fromTask(T.of(1))({})()
-    assert.deepStrictEqual(e, 1)
+    assert.deepStrictEqual(await _.fromTask(T.of(1))({})(), 1)
   })
 
   it('fromReader', async () => {
-    const e = await _.fromReader(R.of(1))({})()
-    assert.deepStrictEqual(e, 1)
+    assert.deepStrictEqual(await _.fromReader(R.of(1))({})(), 1)
   })
 
   // -------------------------------------------------------------------------------------
@@ -92,35 +80,33 @@ describe('ReaderTask', () => {
 
   it('local', async () => {
     const len = (s: string): number => s.length
-    const e = await pipe(
-      _.asks((n: number) => n + 1),
-      _.local(len)
-    )('aaa')()
-    assert.deepStrictEqual(e, 4)
+    assert.deepStrictEqual(
+      await pipe(
+        _.asks((n: number) => n + 1),
+        _.local(len)
+      )('aaa')(),
+      4
+    )
   })
 
   it('chainIOK', async () => {
     const f = (s: string) => I.of(s.length)
-    const x = await pipe(_.of('a'), _.chainIOK(f))(undefined)()
-    assert.deepStrictEqual(x, 1)
+    assert.deepStrictEqual(await pipe(_.of('a'), _.chainIOK(f))(undefined)(), 1)
   })
 
   it('chainTaskK', async () => {
     const f = (s: string) => T.of(s.length)
-    const x = await pipe(_.of('a'), _.chainTaskK(f))(undefined)()
-    assert.deepStrictEqual(x, 1)
+    assert.deepStrictEqual(await pipe(_.of('a'), _.chainTaskK(f))(undefined)(), 1)
   })
 
   it('fromIOK', async () => {
     const f = _.fromIOK((s: string) => I.of(s.length))
-    const x = await pipe(_.of('a'), _.chain(f))({})()
-    assert.deepStrictEqual(x, 1)
+    assert.deepStrictEqual(await pipe(_.of('a'), _.chain(f))({})(), 1)
   })
 
   it('fromTaskK', async () => {
     const f = _.fromTaskK((s: string) => T.of(s.length))
-    const x = await pipe(_.of('a'), _.chain(f))({})()
-    assert.deepStrictEqual(x, 1)
+    assert.deepStrictEqual(await pipe(_.of('a'), _.chain(f))({})(), 1)
   })
 
   // -------------------------------------------------------------------------------------
@@ -129,18 +115,14 @@ describe('ReaderTask', () => {
 
   it('getSemigroup', async () => {
     const M = _.getSemigroup(semigroupString)
-    const e = await M.concat(_.of('a'), _.of('b'))({})()
-    assert.deepStrictEqual(e, 'ab')
+    assert.deepStrictEqual(await M.concat(_.of('a'), _.of('b'))({})(), 'ab')
   })
 
   it('getMonoid', async () => {
     const M = _.getMonoid(monoidString)
-    const e = await M.concat(_.of('a'), M.empty)({})()
-    assert.deepStrictEqual(e, 'a')
-    const e2 = await M.concat(M.empty, _.of('b'))({})()
-    assert.deepStrictEqual(e2, 'b')
-    const e3 = await M.concat(_.of('a'), _.of('b'))({})()
-    assert.deepStrictEqual(e3, 'ab')
+    assert.deepStrictEqual(await M.concat(_.of('a'), M.empty)({})(), 'a')
+    assert.deepStrictEqual(await M.concat(M.empty, _.of('b'))({})(), 'b')
+    assert.deepStrictEqual(await M.concat(_.of('a'), _.of('b'))({})(), 'ab')
   })
 
   it('applicativeTaskEitherSeq', async () => {
@@ -150,6 +132,10 @@ describe('ReaderTask', () => {
   it('applicativeTaskEitherPar', async () => {
     await assertPar(_.ApplicativePar, { fromTask: _.fromTask }, (fa) => fa(null)())
   })
+
+  // -------------------------------------------------------------------------------------
+  // utils
+  // -------------------------------------------------------------------------------------
 
   it('do notation', async () => {
     assert.deepStrictEqual(