+++ test/StateReaderTaskEither.ts
@@ -14,92 +14,86 @@ import * as TE from '../src/TaskEither'
 import * as A from '../src/Array'
 import { assertSeq } from './util'
 
+const state: unknown = {}
+
 describe('StateReaderTaskEither', () => {
   describe('pipeables', () => {
     it('alt', async () => {
-      const e1 = await _.evalState(
-        pipe(
-          _.right('a'),
-          _.alt(() => _.left(1))
-        ),
-        {}
+      const e1 = await pipe(
+        _.right('a'),
+        _.alt(() => _.left(1)),
+        _.evaluate(state)
       )({})()
       assert.deepStrictEqual(e1, E.right('a'))
-      const e2 = await _.evalState(
+      const e2 = await pipe(
         pipe(
           _.left(1),
-          _.alt(() => _.right('b'))
-        ),
-        {}
+          _.alt(() => _.right('b')),
+          _.evaluate(state)
+        )
       )({})()
       assert.deepStrictEqual(e2, E.right('b'))
-      const e3 = await _.evalState(
+      const e3 = await pipe(
         pipe(
           _.left(1),
-          _.alt(() => _.left(2))
-        ),
-        {}
+          _.alt(() => _.left(2)),
+          _.evaluate(state)
+        )
       )({})()
       assert.deepStrictEqual(e3, E.left(2))
     })
 
     it('map', async () => {
       const len = (s: string): number => s.length
-      const e = await _.evalState(pipe(_.right('aaa'), _.map(len)), {})({})()
+      const e = await pipe(_.right('aaa'), _.map(len), _.evaluate(state))({})()
       assert.deepStrictEqual(e, E.right(3))
     })
 
     it('ap', async () => {
       const len = (s: string): number => s.length
-      const e = await _.evalState(pipe(_.right(len), _.ap(_.right('aaa'))), {})({})()
+      const e = await pipe(_.right(len), _.ap(_.right('aaa')), _.evaluate(state))({})()
       assert.deepStrictEqual(e, E.right(3))
     })
 
     it('apFirst', async () => {
-      const e = await _.evalState(pipe(_.right('a'), _.apFirst(_.right('b'))), {})({})()
+      const e = await pipe(_.right('a'), _.apFirst(_.right('b')), _.evaluate(state))({})()
       assert.deepStrictEqual(e, E.right('a'))
     })
 
     it('apSecond', async () => {
-      const e = await _.evalState(pipe(_.right('a'), _.apSecond(_.right('b'))), {})({})()
+      const e = await pipe(_.right('a'), _.apSecond(_.right('b')), _.evaluate(state))({})()
       assert.deepStrictEqual(e, E.right('b'))
     })
 
     it('chain', async () => {
       const f = (s: string) => (s.length > 2 ? _.right(s.length) : _.right(0))
-      const e = await _.evalState(pipe(_.right('aaa'), _.chain(f)), {})({})()
+      const e = await pipe(_.right('aaa'), _.chain(f), _.evaluate(state))({})()
       assert.deepStrictEqual(e, E.right(3))
     })
 
     it('chainFirst', async () => {
       const f = (s: string) => (s.length > 2 ? _.right(s.length) : _.right(0))
-      const e = await _.evalState(pipe(_.right('aaa'), _.chainFirst(f)), {})({})()
+      const e = await pipe(_.right('aaa'), _.chainFirst(f), _.evaluate(state))({})()
       assert.deepStrictEqual(e, E.right('aaa'))
     })
 
-    it('chainFirstW', async () => {
-      const f = (s: string) => (s.length > 2 ? _.right<unknown, unknown, string, number>(s.length) : _.right(0))
-      const e = await _.evalState(pipe(_.right<unknown, unknown, boolean, string>('aaa'), _.chainFirstW(f)), {})({})()
-      assert.deepStrictEqual(e, E.right('aaa'))
-    })
-
-    it('flatten', async () => {
-      const e = await _.evalState(pipe(_.right(_.right('a')), _.flatten), {})({})()
+    it('chainFirst', async () => {
+      const e = await pipe(_.right(_.right('a')), _.flatten, _.evaluate(state))({})()
       assert.deepStrictEqual(e, E.right('a'))
     })
 
     it('bimap', async () => {
       const gt2 = (n: number): boolean => n > 2
       const len = (s: string): number => s.length
-      const e1 = await _.evalState(pipe(_.right('aaa'), _.bimap(gt2, len)), {})({})()
+      const e1 = await pipe(_.right('aaa'), _.bimap(gt2, len), _.evaluate(state))({})()
       assert.deepStrictEqual(e1, E.right(3))
-      const e2 = await _.evalState(pipe(_.left(3), _.bimap(gt2, len)), {})({})()
+      const e2 = await pipe(_.left(3), _.bimap(gt2, len), _.evaluate(state))({})()
       assert.deepStrictEqual(e2, E.left(true))
     })
 
     it('mapLeft', async () => {
       const gt2 = (n: number): boolean => n > 2
-      const e = await _.evalState(pipe(_.left(3), _.mapLeft(gt2)), {})({})()
+      const e = await pipe(_.left(3), _.mapLeft(gt2), _.evaluate(state))({})()
       assert.deepStrictEqual(e, E.left(true))
     })
 
@@ -110,36 +104,32 @@ describe('StateReaderTaskEither', () => {
       const refinement = (u: string | number): u is number => typeof u === 'number'
       const isNumber = _.fromPredicate(refinement, (u) => `Invalid number ${String(u)}`)
 
-      const e1 = await _.evalState(gt2(3), {})({})()
-      const e2 = await _.evalState(gt2(1), {})({})()
-      const e3 = await _.evalState(isNumber(4), {})({})()
+      const e1 = await pipe(gt2(3), _.evaluate(state))({})()
+      const e2 = await pipe(gt2(1), _.evaluate(state))({})()
+      const e3 = await pipe(isNumber(4), _.evaluate(state))({})()
       assert.deepStrictEqual(e1, E.right(3))
       assert.deepStrictEqual(e2, E.left('Invalid number 1'))
       assert.deepStrictEqual(e3, E.right(4))
     })
 
     it('filterOrElse', async () => {
-      const e1 = await _.evalState(
-        pipe(
-          _.right(12),
-          _.filterOrElse(
-            (n) => n > 10,
-            () => 'a'
-          )
+      const e1 = await pipe(
+        _.right(12),
+        _.filterOrElse(
+          (n) => n > 10,
+          () => 'a'
         ),
-        {}
+        _.evaluate(state)
       )({})()
       assert.deepStrictEqual(e1, E.right(12))
 
-      const e2 = await _.evalState(
-        pipe(
-          _.right(8),
-          _.filterOrElse(
-            (n) => n > 10,
-            () => 'a'
-          )
+      const e2 = await pipe(
+        _.right(8),
+        _.filterOrElse(
+          (n) => n > 10,
+          () => 'a'
         ),
-        {}
+        _.evaluate(state)
       )({})()
       assert.deepStrictEqual(e2, E.left('a'))
     })
@@ -181,28 +171,27 @@ describe('StateReaderTaskEither', () => {
     assert.deepStrictEqual(log, ['start 1', 'end 1', 'start 2', 'end 2'])
   })
 
-  it('execState', async () => {
-    const ma = _.right('aaa')
-    const s = {}
-    const e = await _.execState(ma, s)({})()
-    assert.deepStrictEqual(e, E.right(s))
+  it('execute', async () => {
+    const ma = _.right('a')
+    const e = await pipe(ma, _.execute(state))({})()
+    assert.deepStrictEqual(e, E.right({}))
   })
 
   it('rightState', async () => {
-    const state: State<{}, number> = (s) => [1, s]
-    const e = await _.evalState(_.rightState(state), {})({})()
+    const s: State<unknown, number> = (s) => [1, s]
+    const e = await pipe(_.rightState(s), _.evaluate(state))({})()
     assert.deepStrictEqual(e, E.right(1))
   })
 
   it('leftState', async () => {
-    const state: State<{}, number> = (s) => [1, s]
-    const e = await _.evalState(_.leftState(state), {})({})()
+    const s: State<unknown, number> = (s) => [1, s]
+    const e = await pipe(_.leftState(s), _.evaluate(state))({})()
     assert.deepStrictEqual(e, E.left(1))
   })
 
   it('fromReaderTaskEither', async () => {
     const rte: RTE.ReaderTaskEither<{}, string, number> = RTE.right(1)
-    const e = await _.evalState(_.fromReaderTaskEither(rte), {})({})()
+    const e = await pipe(_.fromReaderTaskEither(rte), _.evaluate(state))({})()
     assert.deepStrictEqual(e, E.right(1))
   })